New update remove duplicates
