+++ data/metrics/social/facebook.csv
@@ -581,82 +581,44 @@ date,audience,net_audience_growth,audience_gained,sent_messages_total,published_
 2022-08-03,3673,3,3,0,0,1,217,2,6,2.76,2,0,0,1,3,0,0,0,1,0,0
 2022-08-04,3673,1,1,0,0,2,163,3,1,0.61,1,0,0,0,0,0,0,0,2,0,0
 2022-08-05,3674,1,1,3,3,2,1543,2,55,3.56,14,1,9,19,12,3,0,0,2,0,0
-2022-08-05,3674,1,1,3,3,2,1543,2,55,3.56,14,1,9,19,12,3,0,0,2,0,0
-2022-08-06,3674,0,0,0,0,2,543,4,18,3.31,7,1,1,4,5,0,0,0,1,1,0
 2022-08-06,3674,0,0,0,0,2,543,4,18,3.31,7,1,1,4,5,0,0,0,1,1,0
 2022-08-07,3676,1,1,0,0,1,343,5,17,4.96,1,0,1,12,3,0,0,0,1,0,0
-2022-08-07,3676,1,1,0,0,1,343,5,17,4.96,1,0,1,12,3,0,0,0,1,0,0
-2022-08-08,3677,1,1,0,0,1,399,4,43,10.78,5,0,2,29,7,0,0,0,1,0,0
 2022-08-08,3677,1,1,0,0,1,399,4,43,10.78,5,0,2,29,7,0,0,0,1,0,0
 2022-08-09,3679,2,2,1,1,1,301,30,12,3.99,4,0,0,1,7,1,0,0,1,0,0
-2022-08-09,3679,2,2,1,1,1,301,30,12,3.99,4,0,0,1,7,1,0,0,1,0,0
-2022-08-10,3680,1,1,1,1,2,352,17,13,3.69,6,0,0,2,5,1,0,0,2,0,0
 2022-08-10,3680,1,1,1,1,2,352,17,13,3.69,6,0,0,2,5,1,0,0,2,0,0
 2022-08-11,3684,4,4,1,1,2,1332,7,50,3.75,10,0,3,25,12,1,0,0,2,0,0
-2022-08-11,3684,4,4,1,1,2,1332,7,50,3.75,10,0,3,25,12,1,0,0,2,0,0
-2022-08-12,3729,15,15,2,2,13,5424,12,328,6.05,150,13,16,39,110,2,0,0,7,6,0
 2022-08-12,3729,15,15,2,2,13,5424,12,328,6.05,150,13,16,39,110,2,0,0,7,6,0
 2022-08-13,3735,2,2,1,1,2,2239,7,136,6.07,47,7,3,18,61,1,0,0,1,1,0
-2022-08-13,3735,2,2,1,1,2,2239,7,136,6.07,47,7,3,18,61,1,0,0,1,1,0
-2022-08-14,3735,1,1,0,0,0,701,3,28,3.99,10,0,0,2,16,0,0,0,0,0,0
 2022-08-14,3735,1,1,0,0,0,701,3,28,3.99,10,0,0,2,16,0,0,0,0,0,0
 2022-08-15,3739,2,2,0,0,0,625,5,27,4.32,7,0,1,7,12,0,0,0,0,0,0
-2022-08-15,3739,2,2,0,0,0,625,5,27,4.32,7,0,1,7,12,0,0,0,0,0,0
-2022-08-16,3739,0,0,4,4,1,1317,6,25,1.9,5,0,5,7,8,4,0,0,0,0,1
 2022-08-16,3739,0,0,4,4,1,1317,6,25,1.9,5,0,5,7,8,4,0,0,0,0,1
 2022-08-17,3740,0,0,0,0,0,558,2,8,1.43,2,0,0,4,2,0,0,0,0,0,0
-2022-08-17,3740,0,0,0,0,0,558,2,8,1.43,2,0,0,4,2,0,0,0,0,0,0
-2022-08-18,3740,0,0,0,0,0,232,2,2,0.86,0,0,0,0,2,0,0,0,0,0,0
 2022-08-18,3740,0,0,0,0,0,232,2,2,0.86,0,0,0,0,2,0,0,0,0,0,0
 2022-08-19,3742,1,1,1,1,0,851,3,44,5.17,22,0,6,13,3,1,0,0,0,0,0
-2022-08-19,3742,1,1,1,1,0,851,3,44,5.17,22,0,6,13,3,1,0,0,0,0,0
-2022-08-20,3742,0,0,0,0,0,447,2,18,4.03,11,0,2,4,1,0,0,0,0,0,0
 2022-08-20,3742,0,0,0,0,0,447,2,18,4.03,11,0,2,4,1,0,0,0,0,0,0
 2022-08-21,3742,0,0,0,0,0,146,3,8,5.48,4,0,0,2,2,0,0,0,0,0,0
-2022-08-21,3742,0,0,0,0,0,146,3,8,5.48,4,0,0,2,2,0,0,0,0,0,0
-2022-08-22,3744,1,1,0,0,1,118,1,3,2.54,1,0,0,1,1,0,0,0,1,0,0
 2022-08-22,3744,1,1,0,0,1,118,1,3,2.54,1,0,0,1,1,0,0,0,1,0,0
 2022-08-23,3746,1,1,0,0,1,184,1,6,3.26,1,0,1,2,2,0,0,0,1,0,0
-2022-08-23,3746,1,1,0,0,1,184,1,6,3.26,1,0,1,2,2,0,0,0,1,0,0
-2022-08-24,3746,0,1,0,0,0,125,5,10,8,1,0,1,0,8,0,0,0,0,0,0
 2022-08-24,3746,0,1,0,0,0,125,5,10,8.0,1,0,1,0,8,0,0,0,0,0,0
 2022-08-25,3747,0,0,0,0,0,76,0,3,3.95,2,0,0,1,0,0,0,0,0,0,0
-2022-08-25,3747,0,0,0,0,0,76,0,3,3.95,2,0,0,1,0,0,0,0,0,0,0
-2022-08-26,3749,1,1,0,0,0,94,1,1,1.06,0,0,0,1,0,0,0,0,0,0,0
 2022-08-26,3749,1,1,0,0,0,94,1,1,1.06,0,0,0,1,0,0,0,0,0,0,0
 2022-08-27,3749,0,0,0,0,0,86,1,2,2.33,0,0,0,1,1,0,0,0,0,0,0
-2022-08-27,3749,0,0,0,0,0,86,1,2,2.33,0,0,0,1,1,0,0,0,0,0,0
 2022-08-28,3749,0,0,0,0,0,137,6,1,0.73,0,0,0,1,0,0,0,0,0,0,0
-2022-08-28,3749,0,0,0,0,0,137,6,1,0.73,0,0,0,1,0,0,0,0,0,0,0
-2022-08-29,3754,1,1,0,0,0,130,0,9,6.92,0,0,0,2,7,0,0,0,0,0,0
 2022-08-29,3754,1,1,0,0,0,130,0,9,6.92,0,0,0,2,7,0,0,0,0,0,0
 2022-08-30,3757,0,1,0,0,1,77,0,0,0.0,0,0,0,0,0,0,0,0,1,0,0
-2022-08-30,3757,0,1,0,0,1,77,0,0,0,0,0,0,0,0,0,0,0,1,0,0
-2022-08-31,3761,1,1,0,0,1,88,3,1,1.14,1,0,0,0,0,0,0,0,1,0,0
 2022-08-31,3761,1,1,0,0,1,88,3,1,1.14,1,0,0,0,0,0,0,0,1,0,0
 2022-09-01,3767,1,1,0,0,0,232,2,8,3.45,0,0,0,4,4,0,0,0,0,0,0
-2022-09-01,3767,1,1,0,0,0,232,2,8,3.45,0,0,0,4,4,0,0,0,0,0,0
 2022-09-02,3768,1,1,0,0,0,85,1,10,11.76,4,0,0,0,6,0,0,0,0,0,0
-2022-09-02,3768,1,1,0,0,0,85,1,10,11.76,4,0,0,0,6,0,0,0,0,0,0
-2022-09-03,3769,1,1,0,0,1,66,6,3,4.55,2,0,0,1,0,0,0,0,1,0,0
 2022-09-03,3769,1,1,0,0,1,66,6,3,4.55,2,0,0,1,0,0,0,0,1,0,0
 2022-09-04,3771,-1,0,0,0,0,33,2,0,0.0,0,0,0,0,0,0,0,0,0,0,0
-2022-09-04,3771,-1,0,0,0,0,33,2,0,0,0,0,0,0,0,0,0,0,0,0,0
-2022-09-05,3773,-1,0,0,0,1,210,1,8,3.81,1,0,1,0,6,0,0,0,1,0,0
 2022-09-05,3773,-1,0,0,0,1,210,1,8,3.81,1,0,1,0,6,0,0,0,1,0,0
 2022-09-06,3774,0,0,0,0,1,143,2,3,2.1,0,0,0,0,3,0,0,0,1,0,0
-2022-09-06,3774,0,0,0,0,1,143,2,3,2.1,0,0,0,0,3,0,0,0,1,0,0
-2022-09-07,3776,1,1,0,0,2,104,7,2,1.92,0,0,0,0,2,0,0,0,2,0,0
 2022-09-07,3776,1,1,0,0,2,104,7,2,1.92,0,0,0,0,2,0,0,0,2,0,0
 2022-09-08,3777,1,1,3,3,0,1148,3,41,3.57,23,0,1,1,16,3,0,0,0,0,0
-2022-09-08,3777,1,1,3,3,0,1148,3,41,3.57,23,0,1,1,16,3,0,0,0,0,0
-2022-09-09,3777,0,0,0,0,0,423,1,13,3.07,9,0,0,0,4,0,0,0,0,0,0
-2022-09-09,3777,0,0,0,0,0,423,1,13,3.07,9,0,0,0,4,0,0,0,0,0,0
-2022-09-10,3777,0,1,0,0,0,116,1,3,2.59,1,0,0,0,2,0,0,0,0,0,0
-2022-09-10,3777,0,1,0,0,0,116,1,3,2.59,1,0,0,0,2,0,0,0,0,0,0
-2022-09-11,3777,0,0,0,0,0,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0
-2022-09-11,3777,0,0,0,0,0,128,0,0,0.0,0,0,0,0,0,0,0,0,0,0,0
-2022-09-12,3776,-1,0,0,0,2,173,2,3,1.73,0,0,0,1,2,0,0,0,2,0,0
-2022-09-12,3776,-1,0,0,0,2,173,2,3,1.73,0,0,0,1,2,0,0,0,2,0,0
-2022-09-13,3793,-1,0,0,0,0,10,0,0,0.0,0,0,0,0,0,0,0,0,0,0,0
-2022-09-13,3793,-1,0,0,0,0,10,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+2022-09-09,3787,0,0,0,0,0,423,1,13,3.07,9,0,0,0,4,0,0,0,0,0,0
+2022-09-10,3787,0,1,0,0,0,116,1,3,2.59,1,0,0,0,2,0,0,0,0,0,0
+2022-09-11,3787,0,0,0,0,0,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0
+2022-09-12,3786,-1,0,0,0,2,173,2,3,1.73,0,0,0,1,2,0,0,0,2,0,0
+2022-09-13,3785,-1,0,0,0,0,57,2,8,14.04,0,0,0,3,5,0,0,0,0,0,0
+2022-09-14,3795,1,1,0,0,1,28,0,1,3.57,0,0,0,0,1,0,0,0,1,0,0
+2022-09-15,3797,0,0,0,0,5,112,0,0,0,0,0,0,0,0,0,0,0,0,0,5
