+++ data/metrics/social/instagram.csv
@@ -581,82 +581,44 @@ date,audience,net_audience_growth,audience_gained,sent_messages_total,published_
 2022-08-03,5167,4,4,0,0,9,39,0,0,0.0,0,0,0,0,0,0,0,3,2,0,3,0
 2022-08-04,5170,3,3,0,0,7,63,0,1,1.59,1,0,0,0,0,0,0,1,3,0,3,0
 2022-08-05,5154,1,3,1,1,10,545,0,1,0.18,1,0,0,1,0,0,0,4,4,0,1,0
-2022-08-05,5154,1,3,1,1,10,545,0,1,0.18,1,0,0,1,0,0,0,4,4,0,1,0
-2022-08-06,5159,5,5,0,0,2,218,0,23,10.55,23,0,0,0,0,0,0,0,2,0,0,0
 2022-08-06,5159,5,5,0,0,2,218,0,23,10.55,23,0,0,0,0,0,0,0,2,0,0,0
 2022-08-07,5164,5,7,0,0,2,127,0,11,8.66,11,0,0,0,0,0,0,1,1,0,0,0
-2022-08-07,5164,5,7,0,0,2,127,0,11,8.66,11,0,0,0,0,0,0,1,1,0,0,0
-2022-08-08,5169,5,5,1,0,3,48,0,4,8.33,4,0,0,0,0,0,0,2,0,1,0,0
 2022-08-08,5169,5,5,1,0,3,48,0,4,8.33,4,0,0,0,0,0,0,2,0,1,0,0
 2022-08-09,5172,3,3,0,0,6,89,0,0,0.0,0,0,0,0,0,0,0,1,3,0,0,0
-2022-08-09,5172,3,3,0,0,6,89,0,0,0,0,0,0,0,0,0,0,1,3,0,0,0
-2022-08-10,5174,2,4,0,0,9,33,0,5,15.15,5,0,0,0,0,0,0,4,2,0,1,0
 2022-08-10,5174,2,4,0,0,9,33,0,5,15.15,5,0,0,0,0,0,0,4,2,0,1,0
 2022-08-11,5178,4,5,11,4,7,1457,0,38,2.61,36,0,2,1,3,0,0,0,2,7,4,0
-2022-08-11,5178,4,5,11,4,7,1457,0,38,2.61,36,0,2,1,3,0,0,0,2,7,4,0
-2022-08-12,5208,30,33,13,2,16,3916,0,229,5.85,225,3,1,1,1,1,2,4,1,10,6,0
 2022-08-12,5208,30,33,13,2,16,3916,0,229,5.85,225,3,1,1,1,1,2,4,1,10,6,0
 2022-08-13,5213,5,5,0,0,6,776,0,5,0.64,5,0,0,0,0,0,1,0,2,0,0,0
-2022-08-13,5213,5,5,0,0,6,776,0,5,0.64,5,0,0,0,0,0,1,0,2,0,0,0
-2022-08-14,5214,1,2,0,0,2,217,0,33,15.21,32,1,0,0,0,0,0,0,1,0,1,0
 2022-08-14,5214,1,2,0,0,2,217,0,33,15.21,32,1,0,0,0,0,0,0,1,0,1,0
 2022-08-15,5218,4,10,0,0,4,117,0,24,20.51,24,0,0,0,0,0,0,3,0,0,1,0
-2022-08-15,5218,4,10,0,0,4,117,0,24,20.51,24,0,0,0,0,0,0,3,0,0,1,0
-2022-08-16,5220,2,3,7,1,8,869,0,36,4.14,35,1,0,1,0,0,2,0,2,6,4,0
 2022-08-16,5220,2,3,7,1,8,869,0,36,4.14,35,1,0,1,0,0,2,0,2,6,4,0
 2022-08-17,5225,5,5,3,3,2,835,0,32,3.83,29,2,1,0,3,0,1,1,0,0,0,0
-2022-08-17,5225,5,5,3,3,2,835,0,32,3.83,29,2,1,0,3,0,1,1,0,0,0,0
-2022-08-18,5224,-1,2,0,0,8,98,0,2,2.04,2,0,0,0,0,0,0,3,2,0,2,0
 2022-08-18,5224,-1,2,0,0,8,98,0,2,2.04,2,0,0,0,0,0,0,3,2,0,2,0
 2022-08-19,5227,3,5,1,1,7,652,0,0,0.0,0,0,0,1,0,0,0,2,0,0,1,0
-2022-08-19,5227,3,5,1,1,7,652,0,0,0,0,0,0,1,0,0,0,2,0,0,1,0
-2022-08-20,5228,1,1,0,0,1,218,0,44,20.18,41,0,3,0,0,0,0,0,0,0,0,0
 2022-08-20,5228,1,1,0,0,1,218,0,44,20.18,41,0,3,0,0,0,0,0,0,0,0,0
 2022-08-21,5228,0,0,0,0,2,106,0,11,10.38,11,0,0,0,0,0,0,0,0,0,2,0
-2022-08-21,5228,0,0,0,0,2,106,0,11,10.38,11,0,0,0,0,0,0,0,0,0,2,0
-2022-08-22,5236,8,12,0,0,0,147,0,7,4.76,6,0,1,0,0,0,0,0,0,0,0,0
 2022-08-22,5236,8,12,0,0,0,147,0,7,4.76,6,0,1,0,0,0,0,0,0,0,0,0
 2022-08-23,5244,8,8,6,1,3,687,0,10,1.46,10,0,0,1,0,0,0,0,2,5,1,0
-2022-08-23,5244,8,8,6,1,3,687,0,10,1.46,10,0,0,1,0,0,0,0,2,5,1,0
 2022-08-24,5254,10,12,0,0,1,270,0,18,6.67,18,0,0,0,0,0,0,0,0,0,1,0
-2022-08-24,5254,10,12,0,0,1,270,0,18,6.67,18,0,0,0,0,0,0,0,0,0,1,0
-2022-08-25,5259,5,5,0,0,5,106,0,2,1.89,2,0,0,0,0,0,0,0,4,0,0,0
 2022-08-25,5259,5,5,0,0,5,106,0,2,1.89,2,0,0,0,0,0,0,0,4,0,0,0
 2022-08-26,5261,2,8,0,0,2,158,0,0,0.0,0,0,0,0,0,0,0,0,2,0,0,0
-2022-08-26,5261,2,8,0,0,2,158,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0
-2022-08-27,5264,3,6,0,0,14,70,0,1,1.43,0,0,1,0,0,0,1,4,2,0,7,0
 2022-08-27,5264,3,6,0,0,14,70,0,1,1.43,0,0,1,0,0,0,1,4,2,0,7,0
 2022-08-28,5268,4,5,0,0,0,110,0,8,7.27,7,1,0,0,0,0,0,0,0,0,0,0
-2022-08-28,5268,4,5,0,0,0,110,0,8,7.27,7,1,0,0,0,0,0,0,0,0,0,0
-2022-08-29,5273,5,5,0,0,1,416,0,2,0.48,2,0,0,0,0,0,0,0,0,0,0,0
 2022-08-29,5273,5,5,0,0,1,416,0,2,0.48,2,0,0,0,0,0,0,0,0,0,0,0
 2022-08-30,5276,3,3,0,0,2,48,0,0,0.0,0,0,0,0,0,0,0,0,2,0,0,0
-2022-08-30,5276,3,3,0,0,2,48,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0
 2022-08-31,5280,4,7,0,0,3,59,0,0,0.0,0,0,0,0,0,0,0,2,1,0,0,0
-2022-08-31,5280,4,7,0,0,3,59,0,0,0,0,0,0,0,0,0,0,2,1,0,0,0
-2022-09-01,5284,4,4,0,0,1,23,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0
 2022-09-01,5284,4,4,0,0,1,23,0,0,0.0,0,0,0,0,0,0,0,0,1,0,0,0
-2022-09-02,5287,3,4,0,0,5,11,0,0,0,0,0,0,0,0,0,0,1,4,0,0,0
 2022-09-02,5287,3,4,0,0,5,11,0,0,0.0,0,0,0,0,0,0,0,1,4,0,0,0
-2022-09-03,5291,4,7,0,0,6,37,0,0,0,0,0,0,0,0,0,0,4,2,0,0,0
 2022-09-03,5291,4,7,0,0,6,37,0,0,0.0,0,0,0,0,0,0,0,4,2,0,0,0
 2022-09-04,5292,1,3,0,0,4,44,0,3,6.82,3,0,0,0,0,0,0,2,2,0,0,0
-2022-09-04,5292,1,3,0,0,4,44,0,3,6.82,3,0,0,0,0,0,0,2,2,0,0,0
-2022-09-05,5299,7,10,0,0,10,64,0,1,1.56,1,0,0,0,0,0,0,6,3,0,0,0
 2022-09-05,5299,7,10,0,0,10,64,0,1,1.56,1,0,0,0,0,0,0,6,3,0,0,0
 2022-09-06,5300,1,3,0,0,4,59,0,3,5.08,3,0,0,0,0,0,0,2,1,0,0,0
-2022-09-06,5300,1,3,0,0,4,59,0,3,5.08,3,0,0,0,0,0,0,2,1,0,0,0
-2022-09-07,5309,9,9,0,0,5,84,0,1,1.19,1,0,0,0,0,0,0,2,1,0,1,0
 2022-09-07,5309,9,9,0,0,5,84,0,1,1.19,1,0,0,0,0,0,0,2,1,0,1,0
 2022-09-08,5312,3,6,1,1,4,725,0,18,2.48,18,0,0,1,0,0,0,3,1,0,0,0
-2022-09-08,5312,3,6,1,1,4,725,0,18,2.48,18,0,0,1,0,0,0,3,1,0,0,0
-2022-09-09,5316,4,4,0,0,4,270,0,32,11.85,32,0,0,0,0,0,0,3,1,0,0,0
 2022-09-09,5316,4,4,0,0,4,270,0,32,11.85,32,0,0,0,0,0,0,3,1,0,0,0
 2022-09-10,5320,4,4,0,0,0,90,0,17,18.89,16,0,1,0,0,0,0,0,0,0,0,0
-2022-09-10,5320,4,4,0,0,0,90,0,17,18.89,16,0,1,0,0,0,0,0,0,0,0,0
-2022-09-11,5323,3,6,0,0,0,57,0,2,3.51,2,0,0,0,0,0,0,0,0,0,0,0
-2022-09-11,5323,3,6,0,0,0,57,0,2,3.51,2,0,0,0,0,0,0,0,0,0,0,0
-2022-09-12,5326,3,7,0,0,5,67,0,4,5.97,4,0,0,0,0,0,0,1,1,0,3,0
-2022-09-12,5326,3,7,0,0,5,67,0,4,5.97,4,0,0,0,0,0,0,1,1,0,3,0
-2022-09-13,5326,0,0,0,0,1,109,0,4,3.67,4,0,0,0,0,0,0,0,0,0,1,0
-2022-09-13,5326,0,0,0,0,1,109,0,4,3.67,4,0,0,0,0,0,0,0,0,0,1,0
+2022-09-11,5326,6,6,0,0,0,57,0,2,3.51,2,0,0,0,0,0,0,0,0,0,0,0
+2022-09-12,5333,7,7,0,0,5,67,0,4,5.97,4,0,0,0,0,0,0,1,1,0,3,0
+2022-09-13,5333,0,0,0,0,1,109,0,4,3.67,4,0,0,0,0,0,0,0,0,0,1,0
+2022-09-14,5333,0,0,0,0,6,98,0,0,0,0,0,0,0,0,0,0,5,1,0,0,0
+2022-09-15,5338,5,5,0,0,1,0,0,2,0,2,0,0,0,0,0,0,0,0,0,1,0
