+++ data/metrics/social/linkedin.csv
@@ -580,83 +580,45 @@ date,audience,net_audience_growth,audience_gained,sent_messages_total,published_
 2022-08-02,2199,11,11,0,0,0,253,0,19,7.51,5,0,1,13,0,0,0
 2022-08-03,2211,4,4,0,0,0,314,0,20,6.37,4,0,2,14,0,0,0
 2022-08-04,2214,5,5,0,0,0,189,0,21,11.11,6,0,0,15,0,0,0
-2022-08-05,2221,4,4,0,0,0,140,0,7,5,1,0,0,6,0,0,0
 2022-08-05,2221,4,4,0,0,0,140,0,7,5.0,1,0,0,6,0,0,0
 2022-08-06,2224,2,2,0,0,0,181,0,14,7.73,1,0,0,13,0,0,0
-2022-08-06,2224,2,2,0,0,0,181,0,14,7.73,1,0,0,13,0,0,0
-2022-08-07,2227,6,6,0,0,0,205,0,17,8.29,3,0,0,14,0,0,0
 2022-08-07,2227,6,6,0,0,0,205,0,17,8.29,3,0,0,14,0,0,0
 2022-08-08,2231,6,6,0,0,0,225,0,12,5.33,1,0,0,11,0,0,0
-2022-08-08,2231,6,6,0,0,0,225,0,12,5.33,1,0,0,11,0,0,0
-2022-08-09,2237,7,7,0,0,0,216,0,13,6.02,7,0,1,5,0,0,0
 2022-08-09,2237,7,7,0,0,0,216,0,13,6.02,7,0,1,5,0,0,0
 2022-08-10,2245,9,9,0,0,0,226,0,5,2.21,2,0,0,3,0,0,0
-2022-08-10,2245,9,9,0,0,0,226,0,5,2.21,2,0,0,3,0,0,0
-2022-08-11,2253,6,6,0,0,0,229,0,16,6.99,3,0,0,13,0,0,0
 2022-08-11,2253,6,6,0,0,0,229,0,16,6.99,3,0,0,13,0,0,0
 2022-08-12,2260,9,9,0,0,0,168,0,5,2.98,0,0,0,5,0,0,0
-2022-08-12,2260,9,9,0,0,0,168,0,5,2.98,0,0,0,5,0,0,0
-2022-08-13,2269,2,2,0,0,0,76,0,2,2.63,0,0,0,2,0,0,0
 2022-08-13,2269,2,2,0,0,0,76,0,2,2.63,0,0,0,2,0,0,0
 2022-08-14,2271,9,9,2,2,0,1719,0,168,9.77,44,0,1,123,2,0,0
-2022-08-14,2271,9,9,2,2,0,1719,0,168,9.77,44,0,1,123,2,0,0
-2022-08-15,2282,13,13,0,0,1,1288,0,112,8.7,28,1,1,82,0,1,0
 2022-08-15,2282,13,13,0,0,1,1288,0,112,8.7,28,1,1,82,0,1,0
 2022-08-16,2296,10,10,1,1,0,789,0,65,8.24,25,0,2,38,1,0,0
-2022-08-16,2296,10,10,1,1,0,789,0,65,8.24,25,0,2,38,1,0,0
-2022-08-17,2303,31,31,0,0,0,998,0,79,7.92,27,0,4,48,0,0,0
 2022-08-17,2303,31,31,0,0,0,998,0,79,7.92,27,0,4,48,0,0,0
 2022-08-18,2333,13,13,0,0,0,510,0,71,13.92,16,0,2,53,0,0,0
-2022-08-18,2333,13,13,0,0,0,510,0,71,13.92,16,0,2,53,0,0,0
-2022-08-19,2345,3,3,0,0,0,479,0,27,5.64,10,0,1,16,0,0,0
 2022-08-19,2345,3,3,0,0,0,479,0,27,5.64,10,0,1,16,0,0,0
 2022-08-20,2348,2,2,0,0,0,346,0,22,6.36,8,0,1,13,0,0,0
-2022-08-20,2348,2,2,0,0,0,346,0,22,6.36,8,0,1,13,0,0,0
-2022-08-21,2351,3,3,0,0,0,384,0,26,6.77,9,0,0,17,0,0,0
 2022-08-21,2351,3,3,0,0,0,384,0,26,6.77,9,0,0,17,0,0,0
 2022-08-22,2354,7,7,0,0,0,375,0,17,4.53,9,0,2,6,0,0,0
-2022-08-22,2354,7,7,0,0,0,375,0,17,4.53,9,0,2,6,0,0,0
-2022-08-23,2359,4,4,0,0,1,488,0,37,7.58,10,1,0,26,0,1,0
 2022-08-23,2359,4,4,0,0,1,488,0,37,7.58,10,1,0,26,0,1,0
 2022-08-24,2370,10,10,0,0,0,327,0,21,6.42,5,0,1,15,0,0,0
-2022-08-24,2370,10,10,0,0,0,327,0,21,6.42,5,0,1,15,0,0,0
-2022-08-25,2373,5,5,0,0,0,117,0,11,9.4,3,0,0,8,0,0,0
 2022-08-25,2373,5,5,0,0,0,117,0,11,9.4,3,0,0,8,0,0,0
 2022-08-26,2377,2,2,0,0,0,130,0,8,6.15,3,0,0,5,0,0,0
-2022-08-26,2377,2,2,0,0,0,130,0,8,6.15,3,0,0,5,0,0,0
-2022-08-27,2381,1,1,0,0,0,142,0,4,2.82,0,0,0,4,0,0,0
 2022-08-27,2381,1,1,0,0,0,142,0,4,2.82,0,0,0,4,0,0,0
 2022-08-28,2381,1,1,0,0,0,85,0,12,14.12,0,0,1,11,0,0,0
-2022-08-28,2381,1,1,0,0,0,85,0,12,14.12,0,0,1,11,0,0,0
-2022-08-29,2382,1,1,0,0,0,124,0,6,4.84,0,0,0,6,0,0,0
 2022-08-29,2382,1,1,0,0,0,124,0,6,4.84,0,0,0,6,0,0,0
 2022-08-30,2383,2,2,0,0,0,111,0,8,7.21,1,0,0,7,0,0,0
-2022-08-30,2383,2,2,0,0,0,111,0,8,7.21,1,0,0,7,0,0,0
-2022-08-31,2385,3,3,0,0,0,57,0,5,8.77,1,0,0,4,0,0,0
 2022-08-31,2385,3,3,0,0,0,57,0,5,8.77,1,0,0,4,0,0,0
 2022-09-01,2390,9,9,0,0,0,81,0,6,7.41,5,0,0,1,0,0,0
-2022-09-01,2390,9,9,0,0,0,81,0,6,7.41,5,0,0,1,0,0,0
 2022-09-02,2399,6,6,0,0,0,70,0,5,7.14,3,0,1,1,0,0,0
-2022-09-02,2399,6,6,0,0,0,70,0,5,7.14,3,0,1,1,0,0,0
-2022-09-03,2403,4,4,0,0,0,90,0,9,10,0,0,0,9,0,0,0
 2022-09-03,2403,4,4,0,0,0,90,0,9,10.0,0,0,0,9,0,0,0
 2022-09-04,2407,1,1,0,0,0,41,0,2,4.88,1,0,0,1,0,0,0
-2022-09-04,2407,1,1,0,0,0,41,0,2,4.88,1,0,0,1,0,0,0
-2022-09-05,2408,3,3,0,0,0,91,0,2,2.2,0,0,0,2,0,0,0
 2022-09-05,2408,3,3,0,0,0,91,0,2,2.2,0,0,0,2,0,0,0
 2022-09-06,2412,6,6,0,0,0,51,0,3,5.88,0,0,1,2,0,0,0
-2022-09-06,2412,6,6,0,0,0,51,0,3,5.88,0,0,1,2,0,0,0
-2022-09-07,2419,4,4,0,0,0,103,0,5,4.85,1,0,1,3,0,0,0
 2022-09-07,2419,4,4,0,0,0,103,0,5,4.85,1,0,1,3,0,0,0
 2022-09-08,2421,3,3,0,0,0,45,0,1,2.22,0,0,0,1,0,0,0
-2022-09-08,2421,3,3,0,0,0,45,0,1,2.22,0,0,0,1,0,0,0
-2022-09-09,2424,2,2,0,0,0,46,0,1,2.17,1,0,0,0,0,0,0
 2022-09-09,2424,2,2,0,0,0,46,0,1,2.17,1,0,0,0,0,0,0
 2022-09-10,2426,1,1,0,0,0,29,0,0,0,0,0,0,0,0,0,0
-2022-09-10,2426,1,1,0,0,0,29,0,0,0.0,0,0,0,0,0,0,0
 2022-09-11,2427,2,2,0,0,0,18,0,0,0,0,0,0,0,0,0,0
-2022-09-11,2427,2,2,0,0,0,18,0,0,0.0,0,0,0,0,0,0,0
 2022-09-12,2429,2,2,0,0,0,89,0,8,8.99,1,0,0,7,0,0,0
-2022-09-12,2429,2,2,0,0,0,89,0,8,8.99,1,0,0,7,0,0,0
-2022-09-13,2431,7,7,0,0,0,53,0,4,7.55,0,0,0,4,0,0,0
-2022-09-13,2431,7,7,0,0,0,53,0,4,7.55,0,0,0,4,0,0,0
+2022-09-13,2431,4,4,0,0,0,53,0,4,7.55,0,0,0,4,0,0,0
+2022-09-14,2438,7,7,0,0,0,25,0,1,4,0,0,0,1,0,0,0
+2022-09-15,2440,4,4,0,0,0,30,0,4,13.33,0,0,0,4,0,0,0
