+++ data/metrics/social/twitter.csv
@@ -581,82 +581,44 @@ date,audience,net_audience_growth,audience_gained,sent_messages_total,published_
 2022-08-03,16508,3,3,0,0,16,314,1,4,1.27,1,0,0,2,0,1,0,0,0,0,9,0,0,0,7
 2022-08-04,16510,2,2,2,2,35,543,1,28,5.16,5,1,2,4,0,16,0,0,2,0,20,0,2,2,11
 2022-08-05,16523,13,13,6,6,55,3577,0,154,4.31,37,2,13,12,0,90,3,0,3,0,24,0,12,7,12
-2022-08-05,16523,13,13,6,6,55,3577,0,154,4.31,37,2,13,12,0,90,3,0,3,0,24,0,12,7,12
-2022-08-06,16530,7,7,0,0,26,1221,0,45,3.69,10,1,8,4,0,22,0,0,0,0,5,0,8,3,10
 2022-08-06,16530,7,7,0,0,26,1221,0,45,3.69,10,1,8,4,0,22,0,0,0,0,5,0,8,3,10
 2022-08-07,16537,7,7,0,0,15,358,0,17,4.75,3,0,0,1,0,13,0,0,0,0,2,0,0,0,13
-2022-08-07,16537,7,7,0,0,15,358,0,17,4.75,3,0,0,1,0,13,0,0,0,0,2,0,0,0,13
-2022-08-08,16548,11,11,1,1,21,346,0,12,3.47,5,0,2,3,0,2,1,0,0,0,9,0,2,0,10
 2022-08-08,16548,11,11,1,1,21,346,0,12,3.47,5,0,2,3,0,2,1,0,0,0,9,0,2,0,10
 2022-08-09,16560,12,12,5,5,74,7743,253,167,2.16,65,3,1,6,0,92,1,2,2,0,45,1,6,5,17
-2022-08-09,16560,12,12,5,5,74,7743,253,167,2.16,65,3,1,6,0,92,1,2,2,0,45,1,6,5,17
-2022-08-10,16576,16,16,4,4,63,9554,105,204,2.14,56,3,25,22,0,98,1,1,2,0,15,0,24,7,17
 2022-08-10,16576,16,16,4,4,63,9554,105,204,2.14,56,3,25,22,0,98,1,1,2,0,15,0,24,7,17
 2022-08-11,16585,9,9,5,5,39,6865,9,128,1.86,31,6,10,20,0,61,1,1,3,0,8,0,10,8,13
-2022-08-11,16585,9,9,5,5,39,6865,9,128,1.86,31,6,10,20,0,61,1,1,3,0,8,0,10,8,13
-2022-08-12,16597,12,12,24,24,415,103795,15,3423,3.3,748,23,155,133,0,2364,5,5,14,0,116,0,157,38,104
 2022-08-12,16597,12,12,24,24,415,103795,15,3423,3.3,748,23,155,133,0,2364,5,5,14,0,116,0,157,38,104
 2022-08-13,16695,98,98,0,0,34,8898,4,205,2.3,39,0,4,29,0,133,0,0,0,0,9,1,5,2,17
-2022-08-13,16695,98,98,0,0,34,8898,4,205,2.3,39,0,4,29,0,133,0,0,0,0,9,1,5,2,17
-2022-08-14,16705,10,10,0,0,29,2565,0,68,2.65,4,2,3,12,0,47,0,0,0,0,12,1,3,1,12
 2022-08-14,16705,10,10,0,0,29,2565,0,68,2.65,4,2,3,12,0,47,0,0,0,0,12,1,3,1,12
 2022-08-15,16714,9,9,0,0,30,3587,0,92,2.56,16,1,11,21,0,43,0,0,0,0,9,0,13,0,8
-2022-08-15,16714,9,9,0,0,30,3587,0,92,2.56,16,1,11,21,0,43,0,0,0,0,9,0,13,0,8
-2022-08-16,16721,7,7,8,1,19,3116,0,52,1.67,9,1,6,12,0,24,1,0,0,7,2,5,6,0,6
 2022-08-16,16721,7,7,8,1,19,3116,0,52,1.67,9,1,6,12,0,24,1,0,0,7,2,5,6,0,6
 2022-08-17,16725,4,4,1,1,22,3435,3,70,2.04,9,0,7,15,0,39,1,0,0,0,3,0,7,3,9
-2022-08-17,16725,4,4,1,1,22,3435,3,70,2.04,9,0,7,15,0,39,1,0,0,0,3,0,7,3,9
-2022-08-18,16731,6,6,0,0,25,1480,1,38,2.57,5,0,4,7,0,22,0,0,0,0,3,0,4,0,18
 2022-08-18,16731,6,6,0,0,25,1480,1,38,2.57,5,0,4,7,0,22,0,0,0,0,3,0,4,0,18
 2022-08-19,16745,14,14,1,1,15,2215,0,75,3.39,9,-1,4,17,0,46,1,0,0,0,1,0,4,0,10
-2022-08-19,16745,14,14,1,1,15,2215,0,75,3.39,9,-1,4,17,0,46,1,0,0,0,1,0,4,0,10
-2022-08-20,16755,10,10,0,0,15,1748,0,38,2.17,6,0,3,13,0,16,0,0,0,0,3,0,3,1,8
 2022-08-20,16755,10,10,0,0,15,1748,0,38,2.17,6,0,3,13,0,16,0,0,0,0,3,0,3,1,8
 2022-08-21,16762,7,7,0,0,24,4174,0,230,5.51,10,0,8,7,0,205,0,0,0,0,3,0,8,3,10
-2022-08-21,16762,7,7,0,0,24,4174,0,230,5.51,10,0,8,7,0,205,0,0,0,0,3,0,8,3,10
-2022-08-22,16771,9,9,0,0,22,2974,0,101,3.4,9,0,6,15,0,71,0,0,0,0,4,0,6,1,11
 2022-08-22,16771,9,9,0,0,22,2974,0,101,3.4,9,0,6,15,0,71,0,0,0,0,4,0,6,1,11
 2022-08-23,16779,8,8,1,1,33,3948,0,152,3.85,20,1,18,36,0,77,1,0,0,0,5,0,18,1,9
-2022-08-23,16779,8,8,1,1,33,3948,0,152,3.85,20,1,18,36,0,77,1,0,0,0,5,0,18,1,9
-2022-08-24,16784,5,5,7,5,31,5738,0,147,2.56,37,4,6,15,0,85,1,0,4,2,9,2,5,5,10
 2022-08-24,16784,5,5,7,5,31,5738,0,147,2.56,37,4,6,15,0,85,1,0,4,2,9,2,5,5,10
 2022-08-25,16794,10,10,0,0,29,1730,0,41,2.37,5,0,3,3,0,30,0,0,0,0,12,0,3,1,13
-2022-08-25,16794,10,10,0,0,29,1730,0,41,2.37,5,0,3,3,0,30,0,0,0,0,12,0,3,1,13
-2022-08-26,16799,5,5,3,1,26,2327,1,36,1.55,7,0,0,2,0,27,1,0,0,2,9,3,0,3,11
 2022-08-26,16799,5,5,3,1,26,2327,1,36,1.55,7,0,0,2,0,27,1,0,0,2,9,3,0,3,11
 2022-08-27,16809,10,10,0,0,11,434,0,12,2.76,1,0,0,0,0,11,0,0,0,0,1,0,0,0,10
-2022-08-27,16809,10,10,0,0,11,434,0,12,2.76,1,0,0,0,0,11,0,0,0,0,1,0,0,0,10
-2022-08-28,16817,8,8,0,0,11,310,0,15,4.84,0,0,0,3,0,12,0,0,0,0,2,0,0,0,9
 2022-08-28,16817,8,8,0,0,11,310,0,15,4.84,0,0,0,3,0,12,0,0,0,0,2,0,0,0,9
 2022-08-29,16821,4,4,0,0,16,262,0,5,1.91,0,0,0,1,0,4,0,0,0,0,8,0,0,0,8
-2022-08-29,16821,4,4,0,0,16,262,0,5,1.91,0,0,0,1,0,4,0,0,0,0,8,0,0,0,8
-2022-08-30,16823,2,2,1,1,24,8554,0,67,0.78,9,0,5,2,0,51,0,1,0,0,6,1,4,2,11
 2022-08-30,16823,2,2,1,1,24,8554,0,67,0.78,9,0,5,2,0,51,0,1,0,0,6,1,4,2,11
 2022-08-31,16831,8,8,0,0,21,1164,0,6,0.52,2,0,1,1,0,2,0,0,0,0,5,0,1,2,13
-2022-08-31,16831,8,8,0,0,21,1164,0,6,0.52,2,0,1,1,0,2,0,0,0,0,5,0,1,2,13
-2022-09-01,16836,5,5,0,0,15,355,0,11,3.1,1,0,0,0,0,10,0,0,0,0,6,0,0,1,8
 2022-09-01,16836,5,5,0,0,15,355,0,11,3.1,1,0,0,0,0,10,0,0,0,0,6,0,0,1,8
 2022-09-02,16841,5,5,1,1,14,1321,0,6,0.45,0,0,0,0,0,6,0,1,0,0,8,0,0,2,4
-2022-09-02,16841,5,5,1,1,14,1321,0,6,0.45,0,0,0,0,0,6,0,1,0,0,8,0,0,2,4
-2022-09-03,16841,0,0,0,0,17,377,0,13,3.45,3,0,2,0,0,8,0,0,0,0,7,0,2,1,7
 2022-09-03,16841,0,0,0,0,17,377,0,13,3.45,3,0,2,0,0,8,0,0,0,0,7,0,2,1,7
 2022-09-04,16848,7,7,0,0,5,184,0,1,0.54,-1,0,0,0,0,2,0,0,0,0,2,0,0,0,3
-2022-09-04,16848,7,7,0,0,5,184,0,1,0.54,-1,0,0,0,0,2,0,0,0,0,2,0,0,0,3
-2022-09-05,16850,2,2,0,0,11,191,0,0,0,0,0,0,0,0,0,0,0,0,0,3,1,0,0,7
 2022-09-05,16850,2,2,0,0,11,191,0,0,0.0,0,0,0,0,0,0,0,0,0,0,3,1,0,0,7
 2022-09-06,16853,3,3,2,2,36,2114,0,59,2.79,29,1,8,12,0,9,1,0,1,0,18,1,8,3,6
-2022-09-06,16853,3,3,2,2,36,2114,0,59,2.79,29,1,8,12,0,9,1,0,1,0,18,1,8,3,6
-2022-09-07,16857,4,4,1,1,30,1784,0,35,1.96,3,0,3,9,0,20,0,1,0,0,16,0,4,2,8
 2022-09-07,16857,4,4,1,1,30,1784,0,35,1.96,3,0,3,9,0,20,0,1,0,0,16,0,4,2,8
 2022-09-08,16863,6,6,1,1,24,3003,0,76,2.53,23,0,4,0,0,49,1,0,0,0,1,0,3,3,17
-2022-09-08,16863,6,6,1,1,24,3003,0,76,2.53,23,0,4,0,0,49,1,0,0,0,1,0,3,3,17
-2022-09-09,16877,14,14,0,0,9,416,1,8,1.92,1,0,0,1,0,6,0,0,0,0,1,0,0,0,8
 2022-09-09,16877,14,14,0,0,9,416,1,8,1.92,1,0,0,1,0,6,0,0,0,0,1,0,0,0,8
 2022-09-10,16883,6,6,0,0,11,189,0,1,0.53,0,0,0,0,0,1,0,0,0,0,2,0,0,0,9
-2022-09-10,16883,6,6,0,0,11,189,0,1,0.53,0,0,0,0,0,1,0,0,0,0,2,0,0,0,9
 2022-09-11,16885,2,2,0,0,6,147,0,2,1.36,1,0,0,1,0,0,0,0,0,0,1,1,0,0,4
-2022-09-11,16885,2,2,0,0,6,147,0,2,1.36,1,0,0,1,0,0,0,0,0,0,1,1,0,0,4
-2022-09-12,16889,4,4,0,0,14,196,0,-3,-1.53,-4,0,0,0,0,1,0,0,0,0,5,1,0,0,8
-2022-09-12,16889,4,4,0,0,14,196,0,-3,-1.53,-4,0,0,0,0,1,0,0,0,0,5,1,0,0,8
-2022-09-13,16895,6,6,0,0,10,81,0,1,1.23,0,0,0,0,0,1,0,0,0,0,4,0,0,0,6
-2022-09-13,16895,6,6,0,0,10,81,0,1,1.23,0,0,0,0,0,1,0,0,0,0,4,0,0,0,6
+2022-09-12,16889,4,4,0,0,14,212,0,-4,-1.89,-5,0,0,0,0,1,0,0,0,0,5,1,0,0,8
+2022-09-13,16895,6,6,0,0,10,131,0,5,3.82,0,0,0,0,0,5,0,0,0,0,4,0,0,0,6
+2022-09-14,16899,4,4,0,0,16,111,0,3,2.7,2,0,1,0,0,0,0,0,0,0,8,1,1,0,6
+2022-09-15,16901,2,2,0,0,13,93,0,3,3.23,0,0,0,0,0,3,0,0,0,0,1,0,0,1,11
