+++ docs/_data/metrics/social_media/all_weekly.csv
@@ -30,10 +30,10 @@ week_starting,audience_last,engagements_total,audience_gained_total,impressions_
 2022-07-11,27148,1221,115,43477
 2022-07-18,27290,1365,147,55927
 2022-07-25,27488,1012,203,29816
-2022-08-01,27604,979,190,26924
-2022-08-08,27925,10704,598,319818
-2022-08-15,28083,2556,322,63240
-2022-08-22,28215,1372,290,42986
-2022-08-29,28318,372,192,28280
-2022-09-05,28412,674,206,23696
-2022-09-12,28445,42,52,1556
+2022-08-01,27604,600,134,17923
+2022-08-08,27925,5352,299,159909
+2022-08-15,28083,1278,161,31620
+2022-08-22,28215,686,145,21493
+2022-08-29,28318,186,96,14140
+2022-09-05,28425,337,103,11848
+2022-09-12,28476,46,46,1388
