+++ docs/_data/metrics/social_media/service/facebook_weekly.csv
@@ -30,10 +30,10 @@ week_starting,audience_last,engagements_total,audience_gained_total,impressions_
 2022-07-11,3634,170,0,3868
 2022-07-18,3648,395,8,9839
 2022-07-25,3667,247,17,8275
-2022-08-01,3676,212,8,6525
-2022-08-08,3735,1220,52,21496
-2022-08-15,3742,264,6,8352
-2022-08-22,3749,52,8,1640
-2022-08-29,3771,62,12,1422
-2022-09-05,3777,140,6,4544
-2022-09-12,3793,6,0,366
+2022-08-01,3676,122,6,4096
+2022-08-08,3735,610,26,10748
+2022-08-15,3742,132,3,4176
+2022-08-22,3749,26,4,820
+2022-08-29,3771,31,6,711
+2022-09-05,3787,70,3,2272
+2022-09-12,3797,12,1,370
