+++ docs/_data/metrics/social_media/service/instagram_weekly.csv
@@ -30,10 +30,10 @@ week_starting,audience_last,engagements_total,audience_gained_total,impressions_
 2022-07-11,5098,11,25,4741
 2022-07-18,5117,19,30,2650
 2022-07-25,5154,44,40,1942
-2022-08-01,5164,87,46,2058
-2022-08-08,5214,628,114,13072
-2022-08-15,5228,298,52,5790
-2022-08-22,5268,92,112,3096
-2022-08-29,5292,10,66,1276
-2022-09-05,5323,148,84,2698
-2022-09-12,5326,16,14,352
+2022-08-01,5164,52,31,1168
+2022-08-08,5214,314,57,6536
+2022-08-15,5228,149,26,2895
+2022-08-22,5268,46,56,1548
+2022-08-29,5292,5,33,638
+2022-09-05,5326,74,42,1349
+2022-09-12,5338,10,12,274
