+++ docs/_data/metrics/social_media/service/linkedin_weekly.csv
@@ -30,10 +30,10 @@ week_starting,audience_last,engagements_total,audience_gained_total,impressions_
 2022-07-11,2122,343,33,4772
 2022-07-18,2150,223,28,4056
 2022-07-25,2187,280,41,3851
-2022-08-01,2227,172,52,2228
-2022-08-08,2271,442,96,5718
-2022-08-15,2351,804,150,9588
-2022-08-22,2381,220,60,3328
-2022-08-29,2407,82,52,1148
-2022-09-05,2427,24,42,766
-2022-09-12,2431,24,18,284
+2022-08-01,2227,134,40,1702
+2022-08-08,2271,221,48,2859
+2022-08-15,2351,402,75,4794
+2022-08-22,2381,110,30,1664
+2022-08-29,2407,41,26,574
+2022-09-05,2427,12,21,383
+2022-09-12,2440,17,17,197
