+++ docs/_data/metrics/social_media/service/twitter_weekly.csv
@@ -30,10 +30,10 @@ week_starting,audience_last,engagements_total,audience_gained_total,impressions_
 2022-07-11,16294,697,57,30096
 2022-07-18,16375,728,81,39382
 2022-07-25,16480,441,105,15748
-2022-08-01,16537,508,84,16113
-2022-08-08,16705,8414,336,279532
-2022-08-15,16762,1190,114,39510
-2022-08-22,16817,1008,110,34922
-2022-08-29,16848,218,62,24434
-2022-09-05,16885,362,74,15688
-2022-09-12,16895,-4,20,554
+2022-08-01,16537,292,57,10957
+2022-08-08,16705,4207,168,139766
+2022-08-15,16762,595,57,19755
+2022-08-22,16817,504,55,17461
+2022-08-29,16848,109,31,12217
+2022-09-05,16885,181,37,7844
+2022-09-12,16901,7,16,547
