+++ scripts/metrics/social/clean.py
@@ -51,7 +51,7 @@ SORT = "date"
 def clean_df_2(df: pd.DataFrame) -> pd.DataFrame:
     df = df.fillna(0)
     makeFloatNullableInt(df)
-    df = df.drop_duplicates()
+    df = df.drop_duplicates(subset=SORT,keep="last")
     df = sort_df(df,SORT)
     return df
